Fix Linux breakage I introduced in r11025. While VC6 compiler was now happy - satisfied that the maximum() call was now unambiguous, Gnu GCC compiler now was confused and found ambiguity.
This now should remove all ambiguity and make everyone happy.
I think the problem lied in the negation -- the negation may have been taking
an int16_t and making it something else.




git-svn-id: http://scm.sipfoundry.org/rep/sipX/main@11028 a612230a-c5fa-0310-af8b-88eea846685b

+++ sipXmediaLib/include/mp/MpDspUtilsSumVect.h
@@ -202,7 +202,7 @@ int MpDspUtils::maxAbs(const int16_t *pSrc, int dataLength)
    for (int i = 1; i < dataLength; i++)
    {
       startValue = maximum(startValue, 
-         ((int16_t)pSrc[i] > 0) ? (int16_t)pSrc[i] : -((int16_t)pSrc[i]));
+         (int16_t)( (pSrc[i] > 0) ? pSrc[i] : -(pSrc[i]) ));
    }
    return startValue;
 }
